requirements.txt
fastapi>=0.110,<1.0
uvicorn[standard]>=0.27,<1.0
celery[redis]>=5.3,<6.0
sqlalchemy>=2.0,<3.0
psycopg2-binary>=2.9,<3.0
pydantic-settings>=2.2,<3.0
python-dotenv>=1.0,<2.0
Pillow>=10.2,<11.0
numpy>=1.26,<2.0
